Change to use hook to wrap the multi node action logic
